added rollup flow to support both import and require style modules
