+++ .eslintrc.json
@@ -1,0 +1,21 @@
+{
+  "env": {
+    "browser": true,
+      "es2021": true,
+      "node": true
+  },
+  "extends": "standard",
+  "overrides": [],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module",
+    "requireConfigFile": false,
+    "babelOptions": {
+      "plugins": ["@babel/plugin-syntax-import-assertions"]
+    }
+  },
+  "rules": {
+    "semi": [2, "always"]
+  },
+  "parser": "@babel/eslint-parser"
+}
