+++ package.json
@@ -2,10 +2,16 @@
   "name": "influence-utils",
   "version": "1.10.4",
   "description": "Utilities to interact with Influence",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "type": "module",
+  "module": "./build/index.js",
+  "exports": {
+    "import": "./build/index.js",
+    "require": "./build/index.cjs"
+  },
   "scripts": {
-    "test": "NODE_ENV=test ./node_modules/.bin/mocha"
+    "test": "NODE_ENV=test ./node_modules/.bin/mocha",
+    "lint": "eslint ./src",
+    "build": "rollup --config ./rollup.config.js"
   },
   "repository": {
     "type": "git",
@@ -18,8 +24,24 @@
   },
   "homepage": "https://github.com/Influenceth/influence-utils#readme",
   "devDependencies": {
+    "@babel/cli": "^7.19.3",
+    "@babel/core": "^7.19.6",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-syntax-import-assertions": "^7.20.0",
+    "@babel/preset-env": "^7.19.4",
+    "@rollup/plugin-babel": "^6.0.2",
+    "acorn-import-assertions": "^1.8.0",
     "chai": "^4.3.6",
-    "mocha": "^10.0.0"
+    "eslint": "^8.26.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-mocha": "^10.1.0",
+    "eslint-plugin-n": "^15.4.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "mocha": "^10.0.0",
+    "rollup-plugin-cleanup": "^3.2.1",
+    "rollup-plugin-import-assert": "^2.1.2",
+    "rollup-plugin-terser": "^7.0.2"
   },
   "dependencies": {
     "starknet": "^4.4.2"
