+++ src/constants.js
@@ -1,17 +1,17 @@
 // Influence global constants
-const MASTER_SEED = 'influence';
-const START_TIMESTAMP = 1609459200; // Zero date timestamp for orbits
+export const MASTER_SEED = 'influence';
+export const START_TIMESTAMP = 1609459200; // Zero date timestamp for orbits
 
 // Asteroid constants and enumerables
-const MAX_RADIUS = 375142; // in meters
-const TOTAL_ASTEROIDS = 250000;
-const REGIONS = [ 'MainBelt', 'Trojans' ];
-const SPECTRAL_TYPES = [ 'C', 'Cm', 'Ci', 'Cs', 'Cms', 'Cis', 'S', 'Sm', 'Si', 'M', 'I' ];
-const RARITIES = [ 'Common', 'Uncommon', 'Rare', 'Superior', 'Exceptional', 'Incomparable' ];
-const SIZES = [ 'Small', 'Medium', 'Large', 'Huge' ];
-const BONUS_MAPS = [
+export const MAX_RADIUS = 375142; // in meters
+export const TOTAL_ASTEROIDS = 250000;
+export const REGIONS = ['MainBelt', 'Trojans'];
+export const SPECTRAL_TYPES = ['C', 'Cm', 'Ci', 'Cs', 'Cms', 'Cis', 'S', 'Sm', 'Si', 'M', 'I'];
+export const RARITIES = ['Common', 'Uncommon', 'Rare', 'Superior', 'Exceptional', 'Incomparable'];
+export const SIZES = ['Small', 'Medium', 'Large', 'Huge'];
+export const BONUS_MAPS = [
   {
-    spectralTypes: [ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 ],
+    spectralTypes: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
     base: { name: 'Yield0', level: 0, modifier: 0, type: 'yield' },
     bonuses: [
       { position: 1, name: 'Yield1', level: 1, modifier: 3, type: 'yield' },
@@ -20,7 +20,7 @@ const BONUS_MAPS = [
     ]
   },
   {
-    spectralTypes: [ 0, 1, 2, 3, 4, 5, 8, 10 ],
+    spectralTypes: [0, 1, 2, 3, 4, 5, 8, 10],
     base: { name: 'Volatile0', level: 0, modifier: 0, type: 'volatile' },
     bonuses: [
       { position: 4, name: 'Volatile1', level: 1, modifier: 10, type: 'volatile' },
@@ -29,7 +29,7 @@ const BONUS_MAPS = [
     ]
   },
   {
-    spectralTypes: [ 1, 3, 4, 5, 6, 7, 8, 9 ],
+    spectralTypes: [1, 3, 4, 5, 6, 7, 8, 9],
     base: { name: 'Metal0', level: 0, modifier: 0, type: 'metal' },
     bonuses: [
       { position: 7, name: 'Metal1', level: 1, modifier: 10, type: 'metal' },
@@ -38,7 +38,7 @@ const BONUS_MAPS = [
     ]
   },
   {
-    spectralTypes: [ 0, 1, 2, 3, 4, 5 ],
+    spectralTypes: [0, 1, 2, 3, 4, 5],
     base: { name: 'Organic0', level: 0, modifier: 0, type: 'organic' },
     bonuses: [
       { position: 10, name: 'Organic1', level: 1, modifier: 10, type: 'organic' },
@@ -47,14 +47,14 @@ const BONUS_MAPS = [
     ]
   },
   {
-    spectralTypes: [ 3, 4, 5, 6, 7, 8 ],
+    spectralTypes: [3, 4, 5, 6, 7, 8],
     base: { name: 'RareEarth0', level: 0, modifier: 0, type: 'rareearth' },
     bonuses: [
       { position: 13, name: 'RareEarth3', level: 3, modifier: 30, type: 'rareearth' }
     ]
   },
   {
-    spectralTypes: [ 1, 3, 4, 5, 6, 7, 8, 9 ],
+    spectralTypes: [1, 3, 4, 5, 6, 7, 8, 9],
     base: { name: 'Fissile0', level: 0, modifier: 0, type: 'fissile' },
     bonuses: [
       { position: 14, name: 'Fissile3', level: 3, modifier: 30, type: 'fissile' }
@@ -63,16 +63,16 @@ const BONUS_MAPS = [
 ];
 
 // Crew member enumerables
-const CREW_COLLECTIONS = [ 'Arvad Specialist', 'Arvad Citizen', 'Arvad Leadership', 'Adalian' ];
-const CREW_CLASSES = [ 'Pilot', 'Engineer', 'Miner', 'Merchant', 'Scientist' ];
-const CREW_CLASS_DESCRIPTIONS = [
+export const CREW_COLLECTIONS = ['Arvad Specialist', 'Arvad Citizen', 'Arvad Leadership', 'Adalian'];
+export const CREW_CLASSES = ['Pilot', 'Engineer', 'Miner', 'Merchant', 'Scientist'];
+export const CREW_CLASS_DESCRIPTIONS = [
   'Often restless, always ready for adventure, pilots are happiest when flying. Their expertise gets them and their crew where they want to go quickly, safely, efficiently.',
   'If there is a problem, an engineer will find an answer. Whether it is buildings, ships, or processes, engineers are thrilled to be the one to solve the puzzle.',
   'On the surface it appears that miners value their strength over their brains, but like their quarry, the reality lies underneath: they are highly skilled and erudite in their area of expertise.',
   'Predicting the ebb and flow of the market is about more than just the raw data, it is about anticipating the needs and desires of humanity- this is what merchants strive to understand.',
   'Motivated by a desire to expand human knowledge, scientists learn from the past, while keeping their gaze steadily on the future.'
 ];
-const CREW_TITLES = [
+export const CREW_TITLES = [
   'Communications Officer', 'Teaching Assistant', 'Librarian', 'Nurse', 'Public Safety Officer', 'Warehouse Worker',
   'Maintenance Technician', 'Systems Administrator', 'Structural Engineer', 'Farmer', 'Line Cook', 'Artist',
   'Block Captain', 'Observatory Technician', 'Teacher', 'Historian', 'Physician Assistant', 'Security Officer',
@@ -83,16 +83,16 @@ const CREW_TITLES = [
   'Physician', 'Intelligence Officer', 'Logistics Manager', 'Facilities Supervisor', 'Systems Architect',
   'Reactor Engineer', 'Plant Geneticist', 'Chef', 'Actor', 'Justice', 'Chief Navigator', 'Provost', 'Chief Archivist',
   'Chief Medical Officer', 'Head of Security', 'Chief Logistics Officer', 'Chief Steward', 'Chief Technology Officer',
-  'Head of Engineering', 'Chief Botanist', 'Chief Cook', 'Entertainment Director', 'High Commander' ];
+  'Head of Engineering', 'Chief Botanist', 'Chief Cook', 'Entertainment Director', 'High Commander'];
 
-const CREW_SEX = [ 'Male', 'Female' ];
-const CREW_OUTFIT = [
+export const CREW_SEX = ['Male', 'Female'];
+export const CREW_OUTFIT = [
   'Light Spacesuit - Blue', 'Light Spacesuit - Purple', 'Light Spacesuit - Orange',
   'Heavy Spacesuit - Red', 'Heavy Spacesuit - Black', 'Heavy Spacesuit - Blue',
   'Lab Coat - White', 'Lab Coat - Yellow', 'Lab Coat - Green',
   'Tool Vest - Orange', 'Tool Vest - Green', 'Tool Vest - Yellow',
   'Jacket - Red', 'Jacket - Green', 'Jacket - Black',
-  'Stationwear - Red',  'Stationwear - Green', 'Stationwear - Black',
+  'Stationwear - Red', 'Stationwear - Green', 'Stationwear - Black',
   'Light Spacesuit - Navigation', 'Stationwear - Education', 'Stationwear - Archival',
   'Lab Coat - Medical', 'Heavy Spacesuit - Security', 'Light Spacesuit - Logistics',
   'Tool Vest - Maintenance', 'Light Spacesuit - Technology', 'Tool Vest - Engineering',
@@ -103,24 +103,24 @@ const CREW_OUTFIT = [
   'Scientist Recruit - Primary', 'Scientist Recruit - Variant'
 ];
 
-const CREW_HAIR = [ 'Bald', 'Mohawk', 'Slickback', 'Curly', 'Buzz', 'Top Knot',
-  'Bun', 'Long', 'Ponytail', 'Pixie', 'Double Bun', 'Shoulder' ];
+export const CREW_HAIR = ['Bald', 'Mohawk', 'Slickback', 'Curly', 'Buzz', 'Top Knot',
+  'Bun', 'Long', 'Ponytail', 'Pixie', 'Double Bun', 'Shoulder'];
 
-const CREW_HAIR_COLOR = [ 'Red', 'Gray', 'Brown', 'Blonde', 'Black' ];
-const CREW_FACIAL_FEATURE = [ 'Scar', 'Piercing', 'Long Beard', 'Full Beard', 'Circle Beard',
-  'Handlebar Mustache', 'Mustache' ];
+export const CREW_HAIR_COLOR = ['Red', 'Gray', 'Brown', 'Blonde', 'Black'];
+export const CREW_FACIAL_FEATURE = ['Scar', 'Piercing', 'Long Beard', 'Full Beard', 'Circle Beard',
+  'Handlebar Mustache', 'Mustache'];
 
-const CREW_HEAD_PIECE = [
+export const CREW_HEAD_PIECE = [
   'Welding Goggles', 'AR Glasses', 'Eyepatch', 'Mask', 'Helmet', 'Navigation Goggles', 'Spectacles',
   'Archival Glasses', 'Medical Glasses', 'Headset', 'Earmuffs', 'Technology Glasses', 'Botany Glasses',
   'Chef Hat', 'Eyepatch - Orange', 'Eyepatch - Gold'
 ];
-const CREW_BONUS_ITEM = [
+export const CREW_BONUS_ITEM = [
   'Glow', 'Drone - Gray', 'Drone - Orange', 'Drone - Green', 'Drone - Yellow', 'Drone - Medical',
   'Drone - Technology', 'Drone - Commander'
 ];
 
-const CREW_TRAITS = [
+export const CREW_TRAITS = [
   {
     name: 'Drive: Survival',
     type: 'cosmetic',
@@ -373,7 +373,7 @@ const CREW_TRAITS = [
   }
 ];
 
-const RESOURCES = {
+export const RESOURCES = {
   1: { name: 'Water' },
   2: { name: 'Hydrogen' },
   3: { name: 'Ammonia' },
@@ -396,323 +396,4 @@ const RESOURCES = {
   20: { name: 'Taenite' },
   21: { name: 'Troilite' },
   22: { name: 'Uraninite' }
-};
-
-/**
- * Returns the bonus information based on its position in the bitpacked bonuses int
- * @param num Position in the bitpacked bonuses int
- */
-const toBonus = (num) => {
-  if (num < 1 || num > 14) return '';
-  let bonus;
-
-  for (let b of BONUS_MAPS) {
-    bonus = b.bonuses.find(p => p.position === num);
-    if (!!bonus) return bonus;
-  }
-};
-
-/**
- * Converts packed bonuses into an array of bonus types including base types
- * @param spectralType The spectral type (int) of the asteroid
- * @param packed The bitpacked bonuses int
- */
-const toBonuses = (packed, spectralType) => {
-  if (spectralType === undefined) throw new Error('Spectral type is required');
-
-  const bonuses = [];
-  let b, p, added;
-
-  for (b of BONUS_MAPS) {
-
-    if (b.spectralTypes.includes(spectralType)) {
-      added = false;
-
-      for (p of b.bonuses) {
-
-        if ((packed & (1 << p.position)) > 0) {
-          bonuses.push(p);
-          added = true;
-        }
-      }
-
-      if (!added) bonuses.push(b.base);
-    }
-  }
-
-  return bonuses;
-};
-
-/**
- * Returns the rarity level of the asteroid based on the bonuses and size
- * @param bonuses
- * @param radius
- */
-const toRarity = (bonuses) => {
-  let rarity = 0;
-
-  for (b of bonuses) {
-    rarity += b.level;
-  }
-
-  if (rarity <= 3) return RARITIES[rarity];
-  if (rarity <= 5) return RARITIES[4];
-  return RARITIES[5];
-};
-
-/**
- * Converts array of relative resources into an object based on the RESOURCES list
- * Result format: { RESOURCE_NAME: value }
- * @param resources array of float values (asc order by resource type)
- */
-const toResources = (resources) => {
-  return resources.reduce((acc, value, index) => {
-    const { name } = RESOURCES[index + 1] || {};
-    if (!name) throw new Error('Invalid index/key');
-    acc[name] = value;
-    return acc;
-  }, {});
-}
-
-/**
- * Returns whether the asteroid has been scanned based on its bitpacked bonuses int
- * @param packed The bitpacked bonuses int
- */
-const isScanned = (packed) => {
-  return ((packed & (1 << 0)) > 0);
-}
-
-/**
- * Returns the spectral type string based on its array value
- * @param num The spectral type int value
- */
-const toSpectralType = (num) => {
-  if (num < 0 || num > 10) return '';
-  return SPECTRAL_TYPES[num];
-};
-
-/**
- * Returns the size string based on the asteroid radius
- * @param rad The asteroid radius int value
- */
-const toSize = (rad) => {
-  if (rad <= 5000) return SIZES[0];
-  if (rad <= 20000) return SIZES[1];
-  if (rad <= 50000) return SIZES[2];
-  return SIZES[3];
-};
-
-/**
- * Returns the collection name the crew member is a part of
- * @param c The unpacked collection id
- */
-const toCrewCollection = (c) => CREW_COLLECTIONS[c - 1];
-
-/**
- * Returns the crew class string based on the unpacked class id
- * @param c The unpacked class id
- */
-const toCrewClass = (c) => CREW_CLASSES[c - 1];
-
-/**
- * Returns the crew class description string based on the unpacked class id
- * @param c The unpacked class id
- */
-const toCrewClassDescription = (c) => CREW_CLASS_DESCRIPTIONS[c - 1];
-
-/**
- * Returns the crew title based on the unpacked title id
- * @param t The unpacked title id
- */
-const toCrewTitle = (t) => CREW_TITLES[t - 1];
-
-const toCrewSex = (s) => CREW_SEX[s - 1];
-const toCrewOutfit = (o) => CREW_OUTFIT[o - 1];
-const toCrewHair = (h) => CREW_HAIR[h];
-const toCrewHairColor = (h) => CREW_HAIR_COLOR[h - 1];
-const toCrewFacialFeature = (f) => CREW_FACIAL_FEATURE[f - 1];
-const toCrewHeadPiece = (h) => CREW_HEAD_PIECE[h - 1];
-const toCrewItem = (i) => CREW_BONUS_ITEM[i - 1];
-const toCrewTrait = (t) => CREW_TRAITS[t - 1];
-
-/**
- * Class that defines an orbit and provides convenience conversion methods
- */
-class KeplerianOrbit {
-  constructor(elements) {
-    this.a = elements.a; // Semi-major axis
-    this.e = elements.e; // Eccentricity
-    this.i = elements.i; // Inclination
-    this.o = elements.o; // Longitude of ascending node
-    this.w = elements.w; // Argument of periapsis
-    this.m = elements.m; // Mean anomoly at epoch
-  }
-
-  /**
-    * The distance in AU from center of the ellipse to the object
-    * @param t Angular parameter (in radians)
-    */
-  getRadius(t) {
-    const a = this.a;
-    const e = this.e;
-    return a * (1 - Math.pow(e, 2)) / (1 + e * Math.cos(t));
-  }
-
-  /**
-   * Returns Cartesian coordinates at a specific angular parameter
-   * @param t Angular parmeter (in radians)
-   */
-  getPosByAngle(t) {
-    const i = this.i;
-    const o = this.o;
-    const w = this.w;
-
-    // Distance to the point from the orbit focus.
-    const r = this.getRadius(t);
-
-    // Cartesian transformation
-    const x = r * (Math.cos(o) * Math.cos(t + w) - Math.sin(o) * Math.sin(t + w) * Math.cos(i));
-    const y = r * (Math.sin(o) * Math.cos(t + w) + Math.cos(o) * Math.sin(t + w) * Math.cos(i));
-    const z = r * (Math.sin(t + w) * Math.sin(i));
-
-    const point = { x: x, y: y, z: z };
-    return point;
-  }
-
-  /**
-   * Returns an numPoints sized array of uniformly (in radians) separated points along the orbit path
-   * @param numPoints Number of points to create along an orbit
-   */
-  getSmoothOrbit(numPoints) {
-    const points = [];
-    const delta = 2 * Math.PI / numPoints;
-    let angle = 0;
-
-    for (let i = 0; i < numPoints; i++) {
-      points.push(this.getPosByAngle(angle));
-      angle += delta;
-    }
-
-    return points;
-  }
-
-  /**
-   * Retrieves the orbital period in days
-   */
-  getPeriod() {
-    const thirdLaw = 0.000007495; // R^3 / T^2
-    return Math.sqrt(Math.pow(this.a, 3) / thirdLaw);
-  }
-
-  /**
-   * Retrieves Cartesian coordinates in AU at a specified elapsed time
-   * @param elapsed Time in days (in-game) since game START_TIMESTAMP
-   */
-  getPositionAtTime(elapsed) {
-    const a = this.a;
-    const e = this.e;
-    const i = this.i;
-    const o = this.o;
-    const w = this.w;
-    const m = this.m;
-
-    // Calculate the longitude of perihelion
-    const p = w + o;
-
-    // Calculate mean motion based on assumption that mass of asteroid <<< Sun
-    const k = 0.01720209895; // Gaussian constant (units are days and AU)
-    const n = k / Math.sqrt(Math.pow(a, 3)); // Mean motion
-
-    // Calcualate the mean anomoly at elapsed time
-    const M = m + (n * elapsed);
-
-    // Estimate the eccentric and true anomolies using an iterative approximation
-    let E1;
-    let E = M;
-    let lastDiff = 1;
-
-    while (lastDiff > 0.0000001) {
-      E1 = M + (e * Math.sin(E));
-      lastDiff = Math.abs(E1 - E);
-      E = E1;
-    }
-
-    // Calculate in heliocentric polar and then convert to cartesian
-    const v = 2 * Math.atan(Math.sqrt((1 + e) / (1 - e)) * Math.tan(E / 2));
-    const r = a * (1 - Math.pow(e, 2)) / (1 + e * Math.cos(v)); // Current radius in AU
-
-    const pos = {
-      x: r * (Math.cos(o) * Math.cos(v + p - o) - (Math.sin(o) * Math.sin(v + p - o) * Math.cos(i))),
-      y: r * (Math.sin(o) * Math.cos(v + p - o) + Math.cos(o) * Math.sin(v + p - o) * Math.cos(i)),
-      z: r * Math.sin(v + p - o) * Math.sin(i)
-    };
-
-    return {
-      x: +pos.x.toFixed(10),
-      y: +pos.y.toFixed(10),
-      z: +pos.z.toFixed(10)
-    }
-  }
-};
-
-/**
- * ABIs for client interaction with Influence
- */
-const ethereumContracts = require('./contracts_ethereum.json');
-const starknetContracts = require('./contracts_starknet.json');
-
-/**
- * Other utils
- */
-const Address = require('./lib/address');
-const Merkle = require('./lib/merkle');
-
-module.exports = {
-  MASTER_SEED,
-  START_TIMESTAMP,
-  MAX_RADIUS,
-  TOTAL_ASTEROIDS,
-  REGIONS,
-  SPECTRAL_TYPES,
-  RARITIES,
-  RESOURCES,
-  SIZES,
-  BONUS_MAPS,
-  CREW_COLLECTIONS,
-  CREW_CLASSES,
-  CREW_CLASS_DESCRIPTIONS,
-  CREW_TITLES,
-  CREW_SEX,
-  CREW_OUTFIT,
-  CREW_HAIR,
-  CREW_HAIR_COLOR,
-  CREW_FACIAL_FEATURE,
-  CREW_HEAD_PIECE,
-  CREW_BONUS_ITEM,
-  CREW_TRAITS,
-  toBonus,
-  toBonuses,
-  toRarity,
-  toResources,
-  isScanned,
-  toSpectralType,
-  toSize,
-  toCrewCollection,
-  toCrewClass,
-  toCrewClassDescription,
-  toCrewTitle,
-  toCrewSex,
-  toCrewOutfit,
-  toCrewHair,
-  toCrewHairColor,
-  toCrewFacialFeature,
-  toCrewHeadPiece,
-  toCrewItem,
-  toCrewTrait,
-  Address,
-  Merkle,
-  KeplerianOrbit,
-  contracts: ethereumContracts, // (for backward compatibility)
-  ethereumContracts,
-  starknetContracts
 };
