+++ test/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "plugins": [
+    "mocha"
+  ],
+  "extends": [
+    "plugin:mocha/recommended"
+  ]
+}
